Move join_lines to ra_ide_api
